requirements.txt
Flask~=0.12
flask-cors>=3.0.2
pyaml
